+++ src/shared/FoodCard/index.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import {
   Text,
   View,
@@ -47,12 +48,6 @@ const FoodCard: React.FC<Props> = ({item, color}) => {
   } else {
     expiry_percentage = (daysSpent / (daysSpent + remainder)) * 100;
   }
-  // console.log({
-  //   daysSpent: daysSpent,
-  //   remainingDays: remainingDays,
-  //   remainder: remainder,
-  // });
-
   const [calendarVisible, setIsCalendarVisible] = useState(false);
   const [editModalVisible, setEditModalVisible] = useState(false);
   const [trashModalVisible, setTrashModalVisible] = useState<boolean>(false);
@@ -110,7 +105,6 @@ const FoodCard: React.FC<Props> = ({item, color}) => {
     setIsCalendarVisible(!calendarVisible);
   };
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const openEditModal = (itemId: string) => {
     setEditModalVisible(true);
     setFoodItem(() => {
@@ -139,13 +133,9 @@ const FoodCard: React.FC<Props> = ({item, color}) => {
   const queryClient = useQueryClient();
   const {mutate, isLoading} = useMutation(updateFood, {
     onMutate: async () => {
-      // console.log(storageLocation?.title);
       // Cancel any outgoing refetches (so they don't overwrite our optimistic update)
       await queryClient.cancelQueries(['allfoods']);
       await queryClient.cancelQueries([`${storageLocation.title}`]);
-      // await queryClient.cancelQueries(['Fridge']);
-      // await queryClient.cancelQueries(['Freezer']);
-      // await queryClient.cancelQueries(['Pantry']);
     },
     // If the mutation fails, use the context we returned above
     onError: () => {
@@ -167,9 +157,6 @@ const FoodCard: React.FC<Props> = ({item, color}) => {
     onSettled: () => {
       queryClient.invalidateQueries(['allfoods']);
       queryClient.invalidateQueries([`${storageLocation.title}`]);
-      // queryClient.invalidateQueries(['Fridge']);
-      // queryClient.invalidateQueries(['Pantry']);
-      // queryClient.invalidateQueries(['Freezer']);
     },
   });
 
@@ -292,7 +279,6 @@ const FoodCard: React.FC<Props> = ({item, color}) => {
             disabled={isLoading}
             onPress={() => {
               closeEditModal();
-              // onRequestClose();
             }}>
             <View style={styles.modalContainer}>
               <View style={styles.calendarView}>
@@ -300,7 +286,6 @@ const FoodCard: React.FC<Props> = ({item, color}) => {
                   <Calendar
                     style={styles.calendar}
                     hideExtraDays={true}
-                    // hideDayNames={true}
                     minDate={moment().toLocaleString()}
                     disableAllTouchEventsForDisabledDays={true}
                     onDayPress={day => {
